pyproject.toml
[tool.poetry]
name = "azure-function-distributed-tracing"
version = "0.1.0"
description = ""
authors = ["Jean-Baptiste Ranson <jranson@microsoft.com>", "Tamara Rahimi <tamararahimi@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python =  ">=3.8,<3.10"
azure-functions = "^1.8.0"
azure-functions-durable = "^1.0.0b1"
black = "^22.3.0"
isort = "^5.10.3"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
requests = "^2.26.0"
azure-monitor-opentelemetry-exporter= ">=1.0.0b24,<2.0.0"
azure-core-tracing-opentelemetry= ">=1.0.0b11,<2.0.0"
opentelemetry-instrumentation-logging= ">=0.45b0,<1.0.0"
opentelemetry-instrumentation-aiohttp-client=">=0.45b0,<1.0.0"
opentelemetry-instrumentation-httpx=">=0.45b0,<1.0.0"
opentelemetry-instrumentation-requests=">=0.45b0,<1.0.0"
opentelemetry-instrumentation-urllib3=">=0.45b0,<1.0.0"
opentelemetry-instrumentation-fastapi=">=0.45b0,<1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
